+++ frontend/src/mixin.js
@@ -30,7 +30,8 @@ export default {
             ],
 
             messages: {
-                errorMissingRequiredGroup: 'This player is not a member of one of the required groups.',
+                errorRequiredForbiddenGroup: 'This player is not a member of any of the required groups or a member' +
+                    ' of one of the forbidden groups.',
                 itemNameAllowedCharsHelp: 'Allowed characters (no spaces): A-Z a-z 0-9 - . _',
             },